+++ packages/create-twilio-function/src/create-twilio-function/versions.js
@@ -6,8 +6,8 @@ module.exports = {
     '@twilio/runtime-handler'
   ].replace(/[\^~]/, ''),
   twilioRun: pkgJson.dependencies['twilio-run'],
-  node: '12',
-  typescript: '^3.8',
+  node: '18',
+  typescript: '^5.3.3',
   serverlessRuntimeTypes: '^1.1',
   copyfiles: '^2.2.0',
 };