+++ packages/serverless-api/examples/deploy.js
@@ -10,7 +10,7 @@ async function run() {
   const result = await client.deployProject({
     ...config,
     overrideExistingService: true,
-    runtime: 'node12',
+    runtime: 'node18',
     env: {
       HELLO: 'ahoy',
       WORLD: 'welt',