+++ packages/serverless-api/src/streams/logs.ts
@@ -7,7 +7,9 @@ import { LogsConfig } from '../types/logs';
 export class LogsStream extends Readable {
   private _pollingFrequency: number;
   private _pollingCacheSize: number;
-  private _interval: NodeJS.Timeout | undefined;
+  // The builds become flaky if this is set to NodeJS.Timer or number as type because TypeScript sometimes infers the wrong one of the two. This solves this problem:
+  // https://stackoverflow.com/questions/55550096/ts2322-type-timeout-is-not-assignable-to-type-number-when-running-unit-te
+  private _interval: ReturnType<typeof setTimeout> | undefined;
   private _viewedSids: Set<Sid>;
   private _viewedLogs: Array<{ sid: Sid; dateCreated: Date }>;
 
@@ -47,9 +49,9 @@ export class LogsStream extends Readable {
         }
       );
       logs
-        .filter(log => !this._viewedSids.has(log.sid))
+        .filter((log) => !this._viewedSids.has(log.sid))
         .reverse()
-        .forEach(log => {
+        .forEach((log) => {
           this.push(log);
         });
 
@@ -68,28 +70,30 @@ export class LogsStream extends Readable {
       // and new logs by stringifying the sid and the date together.
       const viewedLogsSet = new Set([
         ...this._viewedLogs.map(
-          log => `${log.sid}-${log.dateCreated.toISOString()}`
+          (log) => `${log.sid}-${log.dateCreated.toISOString()}`
         ),
-        ...logs.map(log => `${log.sid}-${log.date_created}`),
+        ...logs.map((log) => `${log.sid}-${log.date_created}`),
       ]);
       // Then we take that set, map over the logs and split them up into sid and
       // date again, sort them most to least recent and chop off the oldest if
       // they are beyond the polling cache size.
       this._viewedLogs = [...viewedLogsSet]
-        .map(logString => {
+        .map((logString) => {
           const [sid, dateCreated] = logString.split('-');
           return { sid, dateCreated: new Date(dateCreated) };
         })
         .sort((a, b) => b.dateCreated.valueOf() - a.dateCreated.valueOf())
         .slice(0, this._pollingCacheSize);
       // Finally we create a set of just SIDs to compare against.
-      this._viewedSids = new Set(this._viewedLogs.map(log => log.sid));
+      this._viewedSids = new Set(this._viewedLogs.map((log) => log.sid));
 
       if (!this.config.tail) {
         this.push(null);
       }
     } catch (err) {
-      this.destroy(err);
+      if (err instanceof Error) {
+        this.destroy(err);
+      }
     }
   }