+++ packages/twilio-run/src/config/deploy.ts
@@ -74,6 +74,7 @@ export async function getConfigFromFlags(
       (externalCliOptions && externalCliOptions.accountSid) ||
       undefined,
     environmentSuffix: flags.environment,
+    region: flags.region,
   });
 
   flags = mergeFlagsAndConfig<DeployCliFlags>(configFlags, flags, cliInfo);
@@ -98,7 +99,8 @@ export async function getConfigFromFlags(
         ? flags.username
         : username.startsWith('AC')
         ? username
-        : externalCliOptions?.accountSid
+        : externalCliOptions?.accountSid,
+      flags.region
     ));
 
   const pkgJson = await readPackageJsonContent(flags);