+++ packages/twilio-run/src/config/list.ts
@@ -63,6 +63,7 @@ export async function getConfigFromFlags(
       (externalCliOptions && externalCliOptions.accountSid) ||
       undefined,
     environmentSuffix: flags.environment,
+    region: flags.region,
   });
 
   flags = mergeFlagsAndConfig<ListCliFlags>(configFlags, flags, cliInfo);
@@ -85,7 +86,8 @@ export async function getConfigFromFlags(
         ? flags.username
         : username.startsWith('AC')
         ? username
-        : externalCliOptions?.accountSid
+        : externalCliOptions?.accountSid,
+      flags.region
     ));
 
   let serviceName = await getServiceNameFromFlags(flags);